+++ src/core/query/index.ts
@@ -6,3 +6,4 @@ export * from "./subgraph";
 export * from "./communities";
 export * from "./projection";
 export * from "./agentSurface";
+export * from "./recipes";